requirements.txt
# Core
python-dotenv>=1.0.0
rich>=13.7.0
prompt_toolkit>=3.0.43

# LLM providers
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0

# MCP client
mcp>=1.0.0

# RAG / embeddings
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Document loading (for RAG ingestion)
beautifulsoup4>=4.12.0
requests>=2.31.0
tiktoken>=0.7.0
